+++ d_ticcmd.h
@@ -27,8 +27,8 @@
 // plus a checksum for internal state consistency.
 typedef struct
 {
-	int8_t	forwardmove;	// *2048 for move
-	int8_t	sidemove;	// *2048 for move
+	int16_t	forwardmove;	// *2048 for move
+	int16_t	sidemove;	// *2048 for move
     int16_t	angleturn;	// <<16 for angle delta
     int16_t	consistancy;	// checks for net game
     byte	chatchar;
